displacement, normals, getVertRotation
Added a random displacement modified that can be applied to a range of
verts. Normals are now calculated by hand and stored in a Dictionary of
Vector3s to match vertices. Added getVertRotation function for deriving
a vert's dir (the direction used when creating the vert - typically
different from the normal with current procedures).