Added addVert, addTriangleIndex to VertexDict
Added addVert function to VertexDict to complement getVert, along with
the removal of 'normalsIndex' from VertexData, as the normals index must
always be the same as the vertices index, which is further the same as
the UVs index. Added addTriangleIndex function to VertexData, which
should later be used to maintain trianglesIndices.