moved mesh building blueprints into MeshShapes
Moved mesh building blueprints into MeshShapes file, in order to improve
organization and enable control over which shapes to build and what args
to use. Modified asset serialization and binary file project settings to
be more github-friendly as the project moves towards demo scenes (so
that scene changes will be easy to track in github).